Switch SPLIT_USB_DETECT loop to uint16_t

+++ quantum/split_common/split_util.c
@@ -58,7 +58,7 @@ volatile bool isLeftHand = true;
 
 #if defined(SPLIT_USB_DETECT)
 static bool usbIsActive(void) {
-    for (uint8_t i = 0; i < (SPLIT_USB_TIMEOUT / SPLIT_USB_TIMEOUT_POLL); i++) {
+    for (uint16_t i = 0; i < (SPLIT_USB_TIMEOUT / SPLIT_USB_TIMEOUT_POLL); i++) {
         // This will return true if a USB connection has been established
         if (usb_connected_state()) {
             return true;
